// hospital/src/app/models/donor.model.ts
export class donor{
    firstName: string;
    middleName: string;
    lastName: string;
    dateOfBirth: Date;
    bloodGroup: string;
    age: number;
    street: string;
    area: string;
    city: string;
    pincode: number;
    emailID: string;
    password: string;
    confirmPassword: string;
    gender: string;
    phoneNumber: number;
}